requirements.txt
appdirs==1.4.3
asgiref==3.2.3
boto3==1.14.46
botocore==1.17.46
certifi==2020.4.5.1
chardet==3.0.4
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.3
django-cors-headers==3.3.0
django-environ==0.4.5
django-rest-framework==0.1.0
django-storages==1.9.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
docutils==0.15.2
environ==1.0
filelock==3.0.12
gunicorn==20.0.4
idna==2.10
jmespath==0.10.0
Pillow==7.1.2
pipenv==2020.6.2
psycopg2==2.8.5
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.24.0
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.0
stripe==2.49.0
urllib3==1.25.10
virtualenv==20.0.4
virtualenv-clone==0.5.4
virtualenvwrapper-win==1.2.6
whitenoise==5.2.0
